requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.0.1
pymongo==4.5.0
Werkzeug==3.0.0